+++ src/act.wizard.c
@@ -12,6 +12,7 @@
 #include "sysdep.h"
 #include "structs.h"
 #include "utils.h"
+#include "password.h"
 #include "spec/spec_binding.h"
 #include "spec/spec_effective_binding.h"
 #include "spec/spec_registry.h"
@@ -4918,10 +4919,12 @@ static int perform_set(struct char_data *ch, struct char_data *vict, int mode, c
       send_to_char(ch, "You cannot change that.\r\n");
       return (0);
     }
-    strncpy(GET_PASSWD(vict), CRYPT(val_arg, GET_NAME(vict)),
-            MAX_PWD_LENGTH); /* strncpy: OK (G_P:MAX_PWD_LENGTH) */
-    *(GET_PASSWD(vict) + MAX_PWD_LENGTH) = '\0';
-    send_to_char(ch, "Password changed to '%s'.\r\n", val_arg);
+    if (!password_hash(val_arg, GET_PASSWD(vict), sizeof(vict->player.passwd)))
+    {
+      send_to_char(ch, "Unable to store that password.\r\n");
+      return (0);
+    }
+    send_to_char(ch, "Password changed.\r\n");
     break;
   case 41: /* poofin */
     if ((vict == ch) || (GET_LEVEL(ch) == LVL_IMPL))
@@ -10734,7 +10737,8 @@ ACMD(do_showwearoff)
 
 ACMD(do_resetpassword)
 {
-  char query[2048], arg1[MAX_NAME_LENGTH], arg2[MAX_PWD_LENGTH], password[MAX_PWD_LENGTH];
+  char query[2048], arg1[MAX_NAME_LENGTH], arg2[MAX_PWD_LENGTH + 1];
+  char password[MAX_PWD_HASH_LENGTH + 1];
   MYSQL_RES *res;
   MYSQL_ROW row;
   bool account_found = false;
@@ -10750,12 +10754,6 @@ ACMD(do_resetpassword)
   if (!*arg2)
   {
     send_to_char(ch, "Please specify what you would like the new password to be.\r\n");
-    return;
-  }
-
-  if (strstr(arg2, ";") || strstr(arg2, "'"))
-  {
-    send_to_char(ch, "Passwords cannot contain ' or ; symbols.\r\n");
     return;
   }
 
@@ -10786,7 +10784,11 @@ ACMD(do_resetpassword)
     return;
   }
 
-  snprintf(password, sizeof(password), "%s", CRYPT(arg2, arg1));
+  if (!password_hash(arg2, password, sizeof(password)))
+  {
+    send_to_char(ch, "Unable to hash that password.  Password not changed.\r\n");
+    return;
+  }
 
   char *escaped_name_update = mysql_escape_string_alloc(conn, arg1);
   char *escaped_password = mysql_escape_string_alloc(conn, password);
@@ -10805,7 +10807,7 @@ ACMD(do_resetpassword)
   free(escaped_password);
   if (!mysql_query(conn, query))
   {
-    send_to_char(ch, "You have updated account %s's password to '%s'.\r\n", arg1, arg2);
+    send_to_char(ch, "You have updated account %s's password.\r\n", arg1);
     return;
   }